feat(open-api-gateway): add package for declarative apis written in OpenAPI
This commit adds a new `open-api-gateway` package, which contains both a projen project type and a CDK construct, designed to make writing REST APIs nice and easy. OpenAPI is used as the core framework for defining apis.

Given an OpenAPI specification file, the CDK construct will create all API gateway resources to deploy the API. Additionally, client and server code is generated to provide type-safe usage and implementation of the API.