feat(open-api-gateway): support larger api specs via s3-based api definition
Rather than inlining the prepared openapi spec in the cloudformation template, we upload the raw
spec as an asset to s3 and prepare it in a custom resource lambda, before writing to s3 again for
api gateway to consume. This increases the size of openapi specs supported since the spec is no
longer included in the template.

BREAKING CHANGE: construct no longer extends SpecRestApi and requires new specPath prop