feat(open-api-gateway): stricter type enforcement for error responses in ts/python lambda handlers
Rather than allowing any error type to be returned by a lambda handler wrapper via a generic
parameter, we enforce that the response body returned corresponds to the correct type for the status
code of the response.

BREAKING CHANGE: lambda handler wrappers no longer accept a type parameter for errors. lambda
handlers must return the matching response body for the status code defined in the specification.