fix(open-api-gateway): make sure references to generated clients are available when spec unchanged
When packages in a monorepo depended on a non-default generated client, synthesis would fail if the
client was not generated (ie the spec didn't change) as the reference to the client was undefined.
Ensure that the projen project for the generated client is always constructed, but the open api
generator is only invoked when necessary.