feat(open-api-gateway): typescript handler input as single object with typed event/context
For better consistency with the python and java interceptors and for easier extensibility in the
future, accept a single object as the request for typescript interceptors. Use @types/aws-lambda to
provide types for event/context instead of any. As with java and python, include an
"interceptorContext" object dedicated for use by interceptors to pass state to other
interceptors/handlers.

BREAKING CHANGE: Typescript handler methods now accept a single object with input, event, context
rather than as separate arguments.