feat(nx-monorepo): support adding custom additional packages to the yarn/npm/pnpm workspace
Allows adding additional packages to the yarn/npm/pnpm workspace via props or the
addWorkspacePackage method. These are added in addition to those automatically added by parenting
them with the monorepo in the .projenrc. This allows you to add packages which may not be controlled
by projen, for example code generated by other tools.

re #177