fix(open-api-gateway): fix python interceptor removal on successive lambda calls
Whenever lambda reused a container to service a request, interceptors were removed for the second
call onwards. This is due to the input interceptors list being created once when declaring the
handler wrapper decorator, but the handler chain build method mutating that list on every call,
being called on each handler invocation. This was only an issue for python interceptors - typescript
and java do not have this behaviour. This change ensures a new sublist is created rather than
mutating the list of interceptors, ensuring that all interceptors are used for every request.

Additionally, this change fixes a broken import in the python sample lambda handler.