fix(open-api-gateway): support primitive types in request and response body in lambda wrappers (#193)
Using primitive types, including 'any', caused the TypeScript and Java lambda handler wrappers to
fail to compile (and the Python one would fail at runtime). Address this by ensuring no
marshalling/demarshalling is performed when the request/response body is a primitive type, treating
it as the raw string received in the api gateway event and returned in the api gateway result. This
also ensures that the request/response body for primitive types is correctly typed as a string -
previously eg an application/pdf request might be incorrectly typed as a 'Blob' without it actually
being coerced into one.