feat(open-api-gateway): upgrade prepare spec lambda to node 18 and suppress s3 nag warning
The PrepareSpec lambda was causing 2 CDK Nag issues when installed in a new app with cdk ^2.53.0.
The latest Node runtime is now v18 so we upgrade the prepare spec lambda to use this. The other CDK
Nag issue was with the prepare spec policy granting access to write to the cdk asset bucket - since
we don't know the hash of the prepared spec at runtime we must include a wildcard in the key. Since
Node v18 is only introduced as of CDK 2.51.0 we upgrade CDK to latest.