feat(open-api-gateway): generate a router for using a single lambda function for all operations
The generated API client now includes a 'handlerRouter' which can be used to route
requests to the appropriate handler if users desire to deploy a single shared lambda for all API
integrations.

In TypeScript, this is type-safe, ensuring all operations are mapped to a handler produced by the
appropriate handler wrapper method.

Additionally we supply a convenience method 'Operations.all' to make it easy to share the same
integration for all operations.

This change includes support for TypeScript and Python - Java coming soon!

re #238