feat(open-api-gateway): java support for router for using single lambda function for all operations
Adds a HandlerRouter class to the generated Java client which can be extended and used as the lambda
entrypoint for all operations. The router requires you to define a method which returns a handler
for every operation. The router can also be decorated with interceptors to apply to all operations.

fix #238