feat(pipeline): compile ts pipeline as part of build by default (#253)
The AwsCdkTypeScriptApp which the PDKPipelineTsProject inherits from disables compilation due to
synth executing via TypeScript anyway. However, any NodejsFunctions that are referenced won't be
type checked which can lead to mistakes. We err on the side of caution by adding TypeScript
compilation to the build by default, given NodejsFunction is used quite frequently.