feat(open-api-gateway): idomatic smithy project structure for better ide support (#280)
This change restructures the smithy-based projects to combine the `model` and `smithy-build`
directories into a single `smithy` directory. This directory has a more similar structure to those
found in the Smithy documentation, and in turn means that IDE tools can be more easily configured to
help with the developer experience.

Specifically with this change, the Smithy IntelliJ plugin can be used to enhance the development
experience when authoring models. https://github.com/iancaffey/smithy-intellij-plugin

We preserve the behaviour of abstracting away the gradle wrapper, whereby it is ignored from the
generated project's source control by default, and copied from the PDK if it does not already
exist.  Similarly we allow users to bring their own `gradle wrapper` by choosing not to ignore
it in source control, and placing it in the `smithy` directory. This replaces the
`gradleWrapperPath` configuration option.

While this is a breaking change, the migration steps are fairly straightforward:

- Upgrade the open-api-gateway package
- `npx projen` to regenerate code - notice the new `smithy` folder in your api project
- Copy the contents of the old `model` directory into `smithy/src/main/smithy`
- If any custom dependencies were added to the `smithy-build/build.gradle` file, add them
to `smithy/build.gradle`
- Delete the old `model` and `smithy-build` directories

BREAKING CHANGE: Moved the location of the smithy model from 'model' to 'smithy/src/main/smithy'.
Removed the gradleWrapperPath configuration option.

fix #278