feat(open-api-gateway): support for multiple tags on operations
This change adds support for tagging operations with multiple tags. This is achieved by using a new
feature of openapi-generator which will only consider the "first" (if any) tag for an operation
during code generation, ensuring that no duplicated code is generated. Note that the "first" tag for
OpenAPI-based projects is the one defined first in the list of tags associated with an operation.
Unfortunately Smithy 1.27.2 and below will sort tags alphabetically in the generated OpenAPI spec,
so the "first" tag is the lowest in the alphabet. This will be addressed in the next Smithy release.

fix #269