feat(type-safe-api): add type safe api, successor to open api gateway
This change introduces Type Safe API, which replaces Open API Gateway going forward. We rename since
with Smithy support, having "Open API" as part of the name is less relevant to users. Type Safe API
emphasises one of the main benefits of the package, type-safety.

The fundamental difference in this new package is that code generation is now a build task, rather
than coupled to project synthesis. This eliminates the need to check generated code in to version
control, meaning smaller and more legible code reviews.

The project is now more modular, with a cleaner separation of the model, infrastructure, clients and
documentation, resulting in a more intuitive directory structure.

Instead of the 6 different projen projects vended by open-api-gateway (which were specific to the
model language (Smithy vs OpenAPI) and infrastructure language (TS, Java or Python), this vends a
single `TypeSafeApiProject` which lets you configure the infrastructure and model language.
Additionally, we expose the `TypeSafeApiModelProject` to allow for  the Smithy/OpenAPI model project
to be used on its own, which addresses #192.

We remove the opinionated "SampleApi" generated for each language, since while this was good for
enabling users to deploy a hello world API with very little code change required, it involved
starting users off in a less ideal way. In Java the whole CDK project was built as a superjar and
deployed as a lambda (much larger jar than you'd really want). In Python the API client was built as
a layer rather than using local dependencies. In typescript we put the handlers in the api infra
project and relied on `NodejsFunction`'s behaviour of looking for both `.ts` and `.js` files so it'd
work as a library or the root cdk app. So instead of setting users up with a less optimal project
structure which "just works" out of the box, we include quick start guides in the README for how to
set up different languages in the right way.

Note that the README can be further extended and broken up into separate files, but should be a good
start for now.

Fix #192