fix(pipeline): remove unneeded cfn exports (#338)
The pipeline package was exporting `CodeRepositoryGRCUrl` and `SonarqubeSecretArn` which meant the
stack could only be deployed once per region. Nothing seems to be importing these exports. After
discussing with the maintainers, we determined that simple outputs are sufficient.

BREAKING CHANGE: `CodeRepositoryGRCUrl` and `SonarqubeSecretArn` are no longer CloudFormation
exports. Any systems that relied on those being exports will need to be updated to look up the
output value by the stack name.

fix #239

Co-authored-by: Adrian Dimech <51220968+agdimech@users.noreply.github.com>