feat(identity): support updated MFA props along with a new UserPoolWithMFA construct
BREAKING CHANGE: UserPoolWithMfa construct has recently been added which provides a more well
rounded set of default configurations for the UserPool. It introduces a breaking change in that
existing customers of the UserIdentity construct prior to v0.15.5 will by default be upgraded to use
the new defaults which will cause deployments to fail. To workaround this, a new cdk context
property ‘@aws-prototyping-sdk/identity:useLegacyMFAProps’ was added to revert back to legacy
defaults.