feat(type-safe-api): context option for synthesizing to sam local compatible template
This adds a new cdk context option `type-safe-api-local` which when set, will use an inline API
definition rather than a reference to S3. When synthesized with this context option, the generated
cloudformation template is compatible with AWS SAM CLI and can be used with the `sam local
start-api` command to run a local development server.

fix #317