feat(type-safe-api): support for api keys (#444)
Adds support for api keys via `apiKeyOptions`. API Keys can be required
when the API key source is 'HEADER'. Keys can be required by default
and/or required/not-required for specific methods.

Note that this change simplifies the way that default authorizers are
applied by applying authorizers only at the method level.