feat(type-safe-api): invoke scripts through package instead of relative paths
Previously scripts were executed with the working directory as the script dir, wherever the
`@aws-prototyping-sdk/type-safe-api` package was installed. The relative path was encoded in
`tasks.json`, which is nice and portable so long as the package is installed somewhere within the
bounds of the project (eg in `node_modules` in the root of the monorepo).

However, when used through JSII (eg defining the `.projenrc` in Java or Python), the package is
installed into a temporary directory, and the relative path becomes machine-specific as it escapes
the bounds of the project.

To address this, we expose the scripts as bins in the type-safe-api package, and invoke them using
`npx`. We can then use the project dir as the working dir for each script, which further simplifies
the script in `tasks.json`. Note that we modify the way in which the script dir is computed within
the scripts to allow for symlinks (ie when a script in bins is installed, it ends up as a symlink,
so to find the script dir we must first resolve the symlink).