fix(cloudscape-website): disable double execution of eslint
react-scripts has a built in eslint plugin which lints the code as part of build without --fix,
which means that users have to manually fix lint errors rather than letting the test task run eslint
with --fix. The react-scripts eslint plugin also lints during the dev task, which is frustrating
while iterating on local changes with the dev server.