feat(type-safe-api): support header parameters in handler wrappers
Include header parameters in the request parameters in the lambda handler wrappers. Previously
headers would appear in the RequestParameters or RequestArrayParameters type but would not be
populated from the API gateway event. Additionally, we support dashes in header parameters by
wrapping the name in quotes (or converting to camelCase for java).