fix(type-safe-api): relax instanceof checks to allow for dependency version mismatches
When the version of nx-monorepo a user used was not the same as the version type-safe-api depends on
(for example a user upgrades type-safe-api without upgrading the monorepo or vice versa), the
standard instanceof check would fail and type-safe-api would run link commands, assuming it is not
part of a repo. This addresses this issue by performing a check based on the class name of the
instance rather than the built-in instanceof check. We follow the prototype chain in case the
instance is a child of the class we are testing for.