feat(type-safe-api): allow users to disable mock data generation (#518)
Add a new option to disable mock data generation (enabled by default). Recently it was discovered
that the mock data generator cannot handle recursive structures, which are valid in Smithy/OpenAPI.
Until we swap out the generator for something that can handle this (possibly by rolling our own), we
allow users to opt-out if they don't need the mocks.