fix(static-website): allow for multiple cloudfront web acls in a single stack (#528)
The custom resource lambda name is unique per stack, but created for every instance of
CloudfrontWebAcl, meaning that multiple instances of CloudfrontWebAcl will conflict with each other
if used within the same stack. This change adds part of the unique construct address hash to the
name to ensure it is unique enough to use multiple instances in the same stack.