refactor(type-safe-api): move generated packages into a generated folder (#533)
This change moves the runtime, infrastructure, libraries and documentation folders into a folder
called generated. This makes it easier for users to distinguish the packages that they should edit
(just the model) vs the ones they should leave alone. This also allows us to add other folders
adjacent to model in the future, for other API components which users may edit.