feat(type-safe-api): generate lambda handler stubs based on model trait
This change adds a new configuration option `handlers` which allows users to opt in to generating
lambda handler projects for java, python, and/or typescript, ready-configured with packaging tasks
for use in a lambda function. Handler stubs are generated for operations annotated with the @handler
trait in Smithy (or the x-handler vendor extension in OpenAPI), only generating the handler stubs
for the chosen language for each operation. Documentation to come once this feature has been further
fleshed out.