feat(pipeline): enable setting CodeBuild properties on FeatureBranch projects
Apply the properties from the PDKPipeline `codeBuildDefaults` property to the CodeBuild project that
creates feature branch pipelines. This enables  configuration such as role policies, filesystems,
and networking. The role policy is useful to enable access to private repos in CodeArtifact.

The feature branch CodeBuild project now defaults to the `STANDARD_7_0` build
image.