feat(type-safe-api): coerce request parameters into defined types for java
Request parameters for java are now of the correct types declared in the model. Note that only
primitive types and single level arrays of primitives are supported (as per Smithy's supported types
for request parameters). OpenAPI allows you to define request parameters as objects with different
encoding strategies but these are not supported by type safe api. Adds validation to make sure that
the request parameters are of the supported types as part of the model build task. Also moves the
Smithy trait injection logic to the model build task such that traits are added to the .api.json
file produced by the model build, rather than as a preprocessing step each time code is generated.

BREAKING CHANGE: Removed RequestArrayParameters from handler inputs and folded them into
RequestParameters. Renamed CorsInterceptor to ResponseHeadersInterceptor. Validated to ensure only
primitives (or arrays of primitives) are used in request paramters (query/path/headers).
RequestParameters coerce string values into their appropriate types for the handler wrappers to
consume, and will respond with a 400 if request parameters are of the wrong type.