fix(type-safe-api): ensure lambda permission policy size does not exceed limit for router
When using the handler router (a single lambda function for all operations), type safe api would
still create a new permission to grant api gateway permissions to invoke the lambda for each
individual operation. With this change, we retain the granular permissions for individual lambdas,
but as soon as a lambda function is used for multiple operations, we replace the granular permission
with a broader permission for api gateway to invoke it for any operation, so as not to exceed the
maximum policy size.

fix #549