fix(type-safe-api): resolve parameter refs prior to parameter validation (#591)
It is legal to use $ref in parameters in an OpenAPI spec, but the validation was assuming these were
not present. Updated the validation to work on a dereferenced clone of the spec. Ensured that the
final parsed spec retains references to ensure the code generator can consolidate models.

Fixes #590