fix(type-safe-api): fix generated react hooks compilation issues
@tanstack/react-query v5 was released earlier, and the hooks project did not pin the version
resulting in code which failed to compile. Additionally the @paginated trait in Smithy specifies the
Smithy member name which is not the name used in the final generated code, since Smithy -> OpenAPI
translation will rename parameters to the name provided in the @httpHeader or @httpQuery traits.
Additionally the generated typescript client will use valid parameter names, eg kebab-case becomes
camelCase, so address this too.

Fixes #604