fix(type-safe-api): fix typescript handler dist path for operations with inline request bodies
When an operation has inline request body schemas (rather than a $ref) in OpenAPI, the operation ID
is suffixed with "Operation" in the generated code, but the operation nickname remains unchanged.
Previously, the operation ID was used for the handler dist reference in the generated functions, but
the nickname was used to generate the actual filename. Address this by using the nickname in both
places for typescript generated functions.

Fixes #606