fix(type-safe-api): fix syntax error for python runtime request parameters with descriptions
Descriptions in Python request parameters would previously render invalid python syntax as quotes
were escaped as &quot;. Use triple curly braces to ensure these are rendered correctly.
Additionally, field references can only be used once per field in a pydantic model, and since fields
with descriptions include the field in the type annotation these would fail at runtime when another
field is used in the assignment (ie foo: Annotated[StrictStr, Field(..., description="...")] =
Field(alias='foo') ). We therefore remove the assignment of a Field value, which is ok since the
alias is not required as the handler wrapper already constructs a dict with the "paramName"s as keys
rather than the "baseName"s which are in the raw request.

Fixes #609