feat(type-safe-api): generate handler tests for java
Generate example junit tests for java handlers. Test generation is tied to handler generation, ie if
the handler already exists the test won't be generated - it is only generated when a handler is
newly generated. This provides a smoother upgrade path as tests won't be generated for existing
projects, and allows users to delete generated tests without worrying about them reappearing on the
next build.

re #570