fix(monorepo-py): remove incorrect typescript pdk dependency for python monorepos
The projen new command adds a dependency on @aws/pdk when creating a new python monorepo, which
causes the initial install to fail as @aws/pdk is not the correct name for the module in pypi. The
workaround is for users to remove this dependency, but we can remove it within the monorepo for a
smoother initial bootstrap experience for python.

Fixes #643