fix(type-safe-api): fix pydantic core missing in lambda issue
The upgrade of the python runtime to the latest openapi generator introduced pydantic v2 which has a
native component (pydantic core). While the appropriate cross-architecture build command is used, a
mismatch in python version can cause this native dependency to appear to be missing at runtime in a
lambda function. This change upgrades the handler project to use python 3.11, ensuring the build
command uses the same version of python as the lambdas.

fix #669