feat(monorepo): align pdk projen peer dep version with synthesized projects projen version
This change ensures all subprojects of a monorepo have the same projen version as the monorepo
itself for non-runtime dependencies, and sets the default projen version to 0.79.x. This also
provides a single place to change when updating the peer dependency version of projen for PDK, which
will automatically ensure vended PDK projects have that same version and we avoid the "unmet peer"
warning in future.

Fixes #661