fix: pdk upgrade command, upgrade-deps for py/java monorepos and add docs
Fix the `pdk upgrade` command to call the projen task rather than the command for the underlying
package manager.

Add `upgrade-deps` tasks for the Python and Java monorepos for consistency, which spawn the
"upgrade" tasks (if present) for subprojects. Also update the `upgrade-deps` task for the TypeScript
monorepo to spawn the "upgrade" task for any non-node project.

Python doesn't have an upgrade task - so we call the "install" task which updates the lockfile.

Added documentation for upgrading dependencies.

Fixes #653