fix(monorepo-py): fix invalid python jsii code generation for syncpack types (#705)
In python, the GroupConfig syncpack interface was generated in the same python file that also
re-exported interfaces which extended it. The interfaces which extended GroupConfig were generated
in separate files (a file for each sub-namespace), and therefore a circular dependency in python was
created. Address this by wrapping the GroupConfig interface in a namespace such that it is generated
in a separate file, meaning the extending interfaces do not depend on the file which re-exports
them.

Fixes #704