feat(type-safe-api): websocket apis for typescript
Add a new `TypeSafeWebSocketApiProject` which can be used to build WebSocket APIs in a similar
manner to REST APIs.

This includes:

* Modelling an API in Smithy or OpenAPI
* Generated CDK infrastructure (WebSocket API and Lambda functions)
* Generated runtimes containing lambda handler wrappers, and server SDK to send messages to clients
* Generated handler stubs with the `@handler` trait
* Generated client and react hooks libraries for use client side
* Integration with infrastructure and website projects
* WebSocket API explorer
* WebSocket API CDK construct with validation schemas as custom resource to scale for larger apis

Only TypeScript is supported currently, but refactoring work has been done to make Python and Java
support more straightforward.

Note the new Java project in `internal/smithy-async-transformer` - this allows us to perform a
transform on users' Smithy models to augment operations with additional traits required for OpenAPI
generation.