fix(type-safe-api): add empty README.md to initial install for generated python projects
The poetry install that runs as part of post synthesis expects there to be some
code present, but code isn't generated until build time. This means that the
first install will fail when either generating the project for the first time or
checking out a fresh copy (since generated code is not checked in to version
control).

To fix this, we add a blank README.md file in addition to the __init__.py file
as the first install step to keep poetry happy until the generator overwrites
it.

Fixes #813