feat(type-safe-api): add commitGeneratedCode option to control generated code
This commit introduces a new `commitGeneratedCode` option to the
TypeSafeApiProject, which allows controlling whether generated code should be
committed or ignored for all generated projects.

The main changes include:

- Add a `commitGeneratedCode` option to the TypeSafeApiProject and related
options interfaces.
- Set the default value of `commitGeneratedCode` to `false`.
- Conditionally add patterns to .gitignore based on the `commitGeneratedCode`
option for all generated projects.
- Update tests to cover the new `commitGeneratedCode` option.

By default, the generated code will be ignored in the repository, except for
Python projects where it will be included to allow for easier distribution and
deployment of the generated artifacts using Poetry.

Fixes: #813