feat(type-safe-api): replace openapi-generator for typescript rest api code generation
This change replaces the java openapi-generator CLI tool with our own code generation tool, which
produces equivalent code. This removes the dependency on Java for the TypeScript REST API projects,
as well as improving the speed of code generation by avoiding package installation steps and the
multiple passes of openapi codegen used to work around limitations of the tool.

Note that this change fixes #789, which means that those hitting the edge case described in the
issue may experience a breaking change.

Generated code packages now covered by the new codegen:
- typescript runtime
- typescript infrastructure
- typescript handlers
- typescript hooks