fix(type-safe-api): path undefined issue for particular node versions
The fs.dirent class uses `parentPath` as the preferred member for the path of a file entry, however
some node versions still use the now deprecated `path` member and `parentPath` is undefined, causing
the generate script to fail to find templates.

Address this by falling back to the `path` member.

Fixes #838