feat(type-safe-api): faster code generation for python infrastructure
Move python infrastructure projects to new codegen.

Note that I've duplicated the ejs templates for the async python infrastructure as it's a useful
starting point for when async support for python is added.

Note that this also fixes #794 by removing the `.lower()` of the full path to the mock json files.