fix(type-safe-api): more robust support for composed schemas in typescript (#856)
Revamped models generated from composed schemas (ie oneOf, anyOf, allOf). Rather than inlining
properties, use intersection and union types to represent allOf and anyOf. oneOf is difficult to
model in typescript, so we opt for the same behaviour as anyOf for now.

OpenAPI generator never worked for anyOf/oneOf with a mixture of object and primitive types, whereas
this change generates valid code for this case.