feat(type-safe-api): add smithy shape library project for sharing smithy models between apis
Adds a `SmithyShapeLibraryProject` which can be used to define Smithy models that can be reused in
different APIs.

Previously, one would have to declare a `SmithyModelProject` to achieve this, but this needed an
ignored `service` to be defined in the model and generated an OpenAPI spec as part of its build,
which is unnecessary.

Fixes #825