fix(type-safe-api): ensure api key auth can be disabled for specific operations
A bug in cloudformation custom resources meant that api key authentication could not be disabled for
individual operations, as the boolean value `false` would be translated into the string `"false"` in
the cloudformation custom resource properties. Since `"false"` is truthy, the security scheme for
api key auth would be added to the operation where it should be disabled.

Fixes #910