fix(static-website): fix deep linking access denied issue and bucket deployment oom kill issue (#921)
CloudFront with S3 OAC returns a 403 instead of a 404 when an object is missing from an S3 bucket,
and so the redirect for single-page apps was broken. Updated to handle both 403 and 404 to be safe.
Additionally bumped the default memory size for the bucket deployment lambda which was being killed
due to exceeding the memory limit for just the default website.

Fixes #915, Fixes #909